+++ api/Cargo.toml
@@ -11,3 +11,7 @@ rand = "0.8"
 serde = { version = "1.0.152", features = ["derive"] }
 log = "0.4.20"
 env_logger = "0.10.0"
+crossbeam = "0.8.2"
+uuid = "1.4.1"
+futures = "0.3.28"
+crossbeam-queue = "0.3.8"
